requirements.txt
PyQt6 
PyOpenGL
numpy 
Pillow
openai
psd-tools